+++ CHANGELOG.md
@@ -1,6 +1,30 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/dusk/compare/v5.0.3...master)
+## [Unreleased](https://github.com/laravel/dusk/compare/v5.3.0...master)
+
+
+## [v5.3.0 (2019-07-11)](https://github.com/laravel/dusk/compare/v5.2.0...v5.3.0)
+
+### Added
+- Add proxy support to the `dusk:install` command ([#659](https://github.com/laravel/dusk/pull/659))
+
+
+## [v5.2.0 (2019-06-25)](https://github.com/laravel/dusk/compare/v5.1.1...v5.2.0)
+
+### Added
+- Add option to fullsize the browser ([#655](https://github.com/laravel/dusk/pull/655))
+
+
+## [v5.1.1 (2019-06-14)](https://github.com/laravel/dusk/compare/v5.1.0...v5.1.1)
+
+### Fixed
+- Add latest version of Facebook Webdriver ([#654](https://github.com/laravel/dusk/pull/654))
+
+
+## [v5.1.0 (2019-05-02)](https://github.com/laravel/dusk/compare/v5.0.3...v5.1.0)
+
+### Added
+- Implement ChromeDriverCommand ([#643](https://github.com/laravel/dusk/pull/643), [60339a5](https://github.com/laravel/dusk/commit/60339a521a1b05e55af7c90b3151557100a0db4d), [#644](https://github.com/laravel/dusk/pull/644))
 
 
 ## [v5.0.3 (2019-04-02)](https://github.com/laravel/dusk/compare/v5.0.2...v5.0.3)